Give userscripts its own module; panic on broken paths
