+++ components/script/dom/htmlheadelement.rs
@@ -3,23 +3,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use dom::bindings::codegen::Bindings::HTMLHeadElementBinding;
-use dom::bindings::codegen::Bindings::NodeBinding::NodeMethods;
-use dom::bindings::codegen::Bindings::DocumentBinding::DocumentMethods;
-use dom::bindings::codegen::InheritTypes::{HTMLElementCast, HTMLHeadElementDerived, NodeCast};
-use dom::bindings::js::{JSRef, OptionalRootable, Temporary, RootedReference};
+use dom::bindings::codegen::InheritTypes::{HTMLElementCast, HTMLHeadElementDerived};
+use dom::bindings::js::{JSRef, Temporary};
 use dom::document::Document;
 use dom::eventtarget::{EventTarget, EventTargetTypeId};
 use dom::element::ElementTypeId;
-use dom::element::AttributeHandlers;
 use dom::htmlelement::{HTMLElement, HTMLElementTypeId};
-use dom::node::{Node, NodeHelpers, NodeTypeId};
+use dom::node::{Node, NodeTypeId};
+use dom::userscripts::load_script;
 use dom::virtualmethods::VirtualMethods;
 use util::str::DOMString;
-use util::opts;
-use util::resource_files::resources_dir_path;
-use std::borrow::ToOwned;
-use std::fs::read_dir;
-use std::path::PathBuf;
 
 #[dom_struct]
 pub struct HTMLHeadElement {
@@ -52,38 +45,6 @@ impl<'a> VirtualMethods for JSRef<'a, HTMLHeadElement> {
         Some(htmlelement as &VirtualMethods)
     }
     fn bind_to_tree(&self, _tree_in_doc: bool) {
-        if let Some(ref path_str) = opts::get().userscripts {
-            let node: &JSRef<Node> = NodeCast::from_borrowed_ref(self);
-            let first_child = node.GetFirstChild().root();
-            let doc = node.owner_doc().root();
-            let doc = doc.r();
-
-            let path = if &**path_str == "" {
-                let mut p = resources_dir_path();
-                p.push("user-agent-js");
-                p
-            } else {
-                PathBuf::new(path_str)
-            };
-
-            let mut files = match read_dir(&path) {
-                Ok(d) => d.filter_map(|e| e.ok()).map(|e| e.path()).collect::<Vec<_>>(),
-                Err(_) => return
-            };
-
-            files.sort();
-
-            for file in files {
-                let name = match file.into_os_string().into_string() {
-                    Ok(ref s) if s.ends_with(".js") => "file://".to_owned() + &s[..],
-                    _ => continue
-                };
-                let new_script = doc.CreateElement("script".to_owned()).unwrap().root();
-                let new_script = new_script.r();
-                new_script.set_string_attribute(&atom!("src"), name);
-                let new_script_node: &JSRef<Node> = NodeCast::from_borrowed_ref(&new_script);
-                node.InsertBefore(*new_script_node, first_child.r()).unwrap();
-            }
-        }
+        load_script(*self);
     }
 }
