+++ components/script/dom/mod.rs
@@ -316,6 +316,7 @@ pub mod treewalker;
 pub mod uievent;
 pub mod urlhelper;
 pub mod urlsearchparams;
+pub mod userscripts;
 pub mod validitystate;
 pub mod virtualmethods;
 pub mod websocket;
